pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quantum-investment-terminal"
version = "0.1.0"
description = "Bloomberg-like investment analysis terminal with Graham-Dodd methodology"
requires-python = ">=3.12"
authors = [
    {name = "Carlos Angarita", email = "carlosdaniloangaritagarcia@gmail.com"}
]
license = {text = "MIT"}

dependencies = [
  # UI Framework
  "PyQt6>=6.7.0",
  "PyQt6-Qt6>=6.7.0",
  "PyQtWebEngine>=6.7.0",
  "pyqtgraph>=0.13.7",
  "qt-material>=2.14",

  # Financial Data
  "yfinance>=0.2.40",
  "pandas-datareader>=0.10.0",
  "httpx>=0.27.0",
  "websockets>=13.0",
  "aiohttp>=3.10.0",

  # Technical Analysis
  "pandas-ta>=0.3.14b",
  "ta>=0.11.0",
  "mplfinance>=0.12.10b",

  # Visualization
  "plotly>=5.24.0",
  "matplotlib>=3.9.0",

  # Portfolio & Risk
  "quantstats>=0.0.62",
  "riskfolio-lib>=6.3.0",
  "PyPortfolioOpt>=1.5.5",
  "vectorbt>=0.26.1",

  # ML & Neural Networks
  "scikit-learn>=1.5.0",
  "lightgbm>=4.5.0",
  "xgboost>=2.1.0",
  "torch>=2.4.0",
  "prophet>=1.1.6",
  "statsmodels>=0.14.3",

  # NLP & Sentiment
  "transformers>=4.44.0",
  "sentence-transformers>=3.1.0",
  "chromadb>=0.5.11",
  "vaderSentiment>=3.3.2",
  "spacy>=3.7.6",

  # PDF Processing
  "pdfplumber>=0.11.4",
  "pypdf>=4.3.1",
  "python-docx>=1.1.2",
  "Pillow>=10.4.0",

  # Database
  "SQLAlchemy>=2.0.35",
  "alembic>=1.13.3",
  "diskcache>=5.6.3",
  "cachetools>=5.5.0",

  # Utilities
  "pydantic>=2.9.0",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "rich>=13.8.0",
  "apscheduler>=3.10.4",
  "tenacity>=9.0.0",
  "tqdm>=4.66.5",
  "numpy>=1.26.0",
  "scipy>=1.14.0",
  "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-qt>=4.4.0",
  "pytest-asyncio>=0.24.0",
  "black>=24.3.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["quantum_terminal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
